# packages/frontend/backend_openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "UsersController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns all users"
          }
        },
        "summary": "Get all users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          }
        },
        "summary": "Create new user",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns user"
          }
        },
        "summary": "Get user by ID",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated"
          }
        },
        "summary": "Update user",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          }
        },
        "summary": "Delete user",
        "tags": [
          "users"
        ]
      }
    }
  },
  "info": {
    "title": "KRUSANT API",
    "description": "API documentation for KRUSANT backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "password123"
          },
          "roles": {
            "description": "User roles",
            "example": [
              "teacher",
              "admin"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "teacherId": {
            "type": "number",
            "description": "Teacher ID if linking to teacher profile",
            "example": 1
          },
          "studentId": {
            "type": "number",
            "description": "Student ID if linking to student profile",
            "example": 1
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "password123"
          },
          "roles": {
            "description": "User roles",
            "example": [
              "teacher",
              "admin"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "teacherId": {
            "type": "number",
            "description": "Teacher ID if linking to teacher profile",
            "example": 1
          },
          "studentId": {
            "type": "number",
            "description": "Student ID if linking to student profile",
            "example": 1
          }
        }
      }
    }
  }
}